admin/requirements.txt
streamlit==1.30.0
requests==2.31.0
pandas==2.1.3
plotly==5.18.0
python-dotenv==1.0.0